moved cursor back into rmf_site_editor
