+++ crates/rmf_site_editor/src/interaction/select_impl/select_anchor.rs
@@ -22,8 +22,7 @@ use bevy::ecs::{hierarchy::ChildOf, schedule::ScheduleConfigs, system::ScheduleS
 use bevy::prelude::*;
 use bevy_impulse::*;
 
-use crate::interaction::{GizmoBlockers, HighlightAnchors};
-use crate::keyboard::KeyboardServices;
+use crate::interaction::{set_visibility, Cursor, GizmoBlockers, HighlightAnchors, IntersectGroundPlaneParams};
 use crate::site::{AnchorBundle, CurrentEditDrawing, DrawingMarker};
 use crate::workspace::CurrentWorkspace;
 use crate::{interaction::select_impl::*, site::CurrentLevel};
