+++ crates/rmf_site_picking/Cargo.toml
@@ -10,6 +10,7 @@ bevy_reflect = {workspace = true}
 bevy_picking = {workspace = true}
 bevy_render = {workspace = true}
 bevy_derive = {workspace = true}
+tracing = {workspace = true}
 bevy_input = {workspace = true}
 bevy_asset = {workspace = true}
 bevy_pbr = {workspace = true}
