--- crates/rmf_site_picking/src/cursor/resources.rs
@@ -1,16 +1,0 @@
-use bevy_asset::Handle;
-use bevy_ecs::prelude::*;
-use bevy_pbr::prelude::*;
-use bevy_render::prelude::*;
-
-#[derive(Resource)]
-pub struct CursorHaloMesh(pub Handle<Mesh>);
-
-#[derive(Resource)]
-pub struct CursorHaloMaterial(pub Handle<StandardMaterial>);
-
-#[derive(Resource)]
-pub struct CursorDaggerMesh(pub Handle<Mesh>);
-
-#[derive(Resource)]
-pub struct CursorDaggerMaterial(pub Handle<StandardMaterial>);
