+++ crates/rmf_site_picking/src/lib.rs
@@ -7,5 +7,5 @@ pub use visual_cue::*;
 pub mod select;
 pub use select::*;
 
-pub mod cursor;
-pub use cursor::*;
+// pub mod cursor;
+// pub use cursor::*;
